[alpha_factory] refactor agents into submodules
